+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_createecdssigningkey.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "eb1f5e65de486aec22eb1b4584cbff56";
+module.exports.hash = "6f192d7cd679142a4ef0136006e203ab";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}