+++ sdk/digitaltwins/digital-twins-core/recordings/node/digitaltwins_components__read_update_and_delete_operations/recording_publish_component_telemetry.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "25ec28d25609550cf93f75f2ee7f66dd";
+module.exports.hash = "da47e556ef0b4cfdf29f95577051ad79";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}