+++ sdk/core/core-tracing/CHANGELOG.md
@@ -6,10 +6,12 @@
 
 ### Breaking Changes
 
+- Removed `TestTracer` and `TestSpan` from public API and into `@azure/test-utils`. [PR #16315](https://github.com/Azure/azure-sdk-for-js/pull/16315)
+  - `TestTracer` and `TestSpan` are intended for test support when used by other Azure packages and not intended for use by end users.
+
 ### Key Bugs Fixed
 
 ### Fixed
-
 
 ## 1.0.0-preview.12 (2021-06-30)